requirements.txt
akshare>=1.9.25
pandas>=1.5.0
numpy>=1.21.0
tqdm>=4.64.0
requests>=2.28.0
PyYAML>=6.0
gitpython>=3.1.30